estudando lógica de programação e calculo de números binários

+++ dio01/src/App.java
@@ -6,9 +6,9 @@ public class App {
         Scanner entrada = new Scanner(System.in);
         System.out.println("Olá, seja bem vindo ao meu programa!");
         System.out.println("Digite um número para ver o seu fatorial: ");
-        byte num = entrada.nextByte();
-        byte numF = num;
-        byte res = 1;
+        int num = entrada.nextByte();
+        int numF = num;
+        int res = 1;
         for(int i= 1; i < num + 1; i++) {
             res *= i;
         }
